requirements.txt
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Limiter==3.8.0
greenlet==3.1.1
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysqlclient==2.2.5
oauthlib==3.2.2
ordered-set==4.1.0
packaging==24.1
Pygments==2.18.0
python-dotenv==1.0.1
rich==13.9.4
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.4
wrapt==1.16.0
